# admin website/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Keto Diet Doctor Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meal-plans:
    get:
      operationId: getLibraryMealPlans
      tags: [mealplans]
      summary: List all library meal plans (doctor-scoped)
      responses:
        "200":
          description: List of library meal plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryMealPlan"
    post:
      operationId: createLibraryMealPlan
      tags: [mealplans]
      summary: Create a new library meal plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryMealPlanRequest"
      responses:
        "201":
          description: Library meal plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryMealPlan"
  /meal-plans/{planId}:
    get:
      operationId: getLibraryMealPlan
      tags: [mealplans]
      summary: Get a library meal plan with its items
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Library meal plan with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryMealPlanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLibraryMealPlan
      tags: [mealplans]
      summary: Update a library meal plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryMealPlanRequest"
      responses:
        "200":
          description: Updated library meal plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryMealPlan"
    delete:
      operationId: deleteLibraryMealPlan
      tags: [mealplans]
      summary: Delete a library meal plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /meal-plans/{planId}/items:
    post:
      operationId: addLibraryMealPlanItem
      tags: [mealplans]
      summary: Add a food item to a library meal plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLibraryMealPlanItemRequest"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryMealPlanItem"
  /meal-plans/{planId}/items/{itemId}:
    delete:
      operationId: deleteLibraryMealPlanItem
      tags: [mealplans]
      summary: Remove an item from a library meal plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /kids/{kidId}/meal-plan:
    get:
      operationId: getKidAssignedMealPlan
      tags: [mealplans]
      summary: Get the library meal plan assigned to a kid (or null if none)
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assigned meal plan with items (or null)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryMealPlanDetail"
        "204":
          description: No plan assigned
    put:
      operationId: assignKidMealPlan
      tags: [mealplans]
      summary: Assign or unassign a library meal plan to a kid
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignMealPlanRequest"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /kids/{kidId}/meal-plans:
    get:
      operationId: getKidMealPlans
      tags: [kids]
      summary: Get all meal plans for a kid
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of meal plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlan"
    post:
      operationId: createKidMealPlan
      tags: [kids]
      summary: Create a new meal plan for a kid
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealPlanRequest"
      responses:
        "201":
          description: Meal plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"
  /kids/{kidId}/meal-plans/{planId}:
    get:
      operationId: getKidMealPlan
      tags: [kids]
      summary: Get a single meal plan with its items
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meal plan with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlanDetail"
    put:
      operationId: updateKidMealPlan
      tags: [kids]
      summary: Update a meal plan
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMealPlanRequest"
      responses:
        "200":
          description: Updated meal plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"
    delete:
      operationId: deleteKidMealPlan
      tags: [kids]
      summary: Delete a meal plan
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /kids/{kidId}/meal-plans/{planId}/items:
    post:
      operationId: addMealPlanItem
      tags: [kids]
      summary: Add a food item to a meal plan
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMealPlanItemRequest"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlanItem"
  /kids/{kidId}/meal-plans/{planId}/items/{itemId}:
    delete:
      operationId: deleteMealPlanItem
      tags: [kids]
      summary: Remove an item from a meal plan
      parameters:
        - name: kidId
          in: path
          required: true
          schema:
            type: integer
        - name: planId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    LibraryMealPlan:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
    LibraryMealPlanItem:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        mealType:
          type: string
        foodName:
          type: string
        portionGrams:
          type: number
        unit:
          type: string
        calories:
          type: number
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        notes:
          type: string
      required:
        - id
        - planId
        - mealType
        - foodName
        - portionGrams
        - unit
    LibraryMealPlanDetail:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryMealPlanItem"
      required:
        - id
        - name
        - items
    CreateLibraryMealPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    UpdateLibraryMealPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    AddLibraryMealPlanItemRequest:
      type: object
      properties:
        mealType:
          type: string
        foodName:
          type: string
        portionGrams:
          type: number
        unit:
          type: string
          default: g
        calories:
          type: number
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        notes:
          type: string
      required:
        - mealType
        - foodName
        - portionGrams
...
    AssignMealPlanRequest:
      type: object
      properties:
        planId:
          type: integer
          nullable: true
      required:
        - planId
    MealPlanItem:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        mealType:
          type: string
        foodId:
          type: integer
        foodName:
          type: string
        portionGrams:
          type: number
        calories:
          type: number
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        notes:
          type: string
      required:
        - id
        - planId
        - mealType
        - foodId
        - foodName
        - portionGrams
    MealPlan:
      type: object
      properties:
        id:
          type: integer
        kidId:
          type: integer
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - kidId
        - name
        - isActive
    MealPlanDetail:
      type: object
      properties:
        id:
          type: integer
        kidId:
          type: integer
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealPlanItem"
      required:
        - id
        - kidId
        - name
        - isActive
        - items
    CreateMealPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
      required:
        - name
    UpdateMealPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    AddMealPlanItemRequest:
      type: object
      properties:
        mealType:
          type: string
        foodId:
          type: integer
        foodName:
          type: string
        portionGrams:
          type: number
        calories:
          type: number
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        notes:
          type: string
      required:
        - mealType
        - foodId
        - foodName
        - portionGrams
...